chore(occt-wasm): extract hull helpers and getNurbsCurveData from adapter
Move 90 lines of convex hull helpers (computeConvexHullFaces,
findHorizonEdges) into hullOps.ts and the optional getNurbsCurveData
into curveOps.ts. Adapter loses ~130 lines and is now free of
module-scope helpers other than wrapKernelExceptions/notImplemented.

Adapter line count: 2416 → 2288 (-128).
